Separated out unit tests to library and library-ui.
